@@ -1,11 +1,7 @@
 import { Routes, Route } from '@angular/router';
 import { TranslateService } from '@ngx-translate/core';
-import { Observable } from 'rxjs/Observable';
-import { Observer } from 'rxjs/Observer';
+import { Observable, Observer } from 'rxjs';
 import { Location } from '@angular/common';
-import 'rxjs/add/observable/forkJoin';
-import 'rxjs/add/operator/share';
-import 'rxjs/add/operator/toPromise';
 import { CacheMechanism, LocalizeRouterSettings } from './localize-router.config';
 import { Inject } from '@angular/core';
 
@@ -110,7 +106,12 @@ export abstract class LocalizeParser {
 
     /** translate routes */
     const res = this.translateRoutes(selectedLanguage);
-    return res.toPromise();
+    return new Promise((resolve, reject) => {
+      res.subscribe(
+        (data) => { resolve(data); },
+        (error) => { reject(error); }
+      );
+    });
   }
 
   initChildRoutes(routes: Routes) {
